+++ src/AAL/process/index.ts
@@ -5,5 +5,6 @@
 export { resolveAgentAction, resolveAgentActionWithScore } from './resolveAgentAction.js'
 export { buildDecisionReason } from './buildDecisionReason.js'
 export type { DecisionReason } from './buildDecisionReason.js'
-export { buildRemovalPlan } from './buildRemovalPlan.js'
+export { buildRemovalPlan, buildRemovalPlanFromResult } from './buildRemovalPlan.js'
 export type { RemovalPlan } from './buildRemovalPlan.js'
+export { applyRemovalPlan } from './applyRemovalPlan.js'